+++ packages/cliapp/src/declarative/mod.ts
@@ -1,0 +1,3 @@
+// Re-export types
+export type * as Option from './option/mod.ts';
+export type { CommandDefinition, InferredOptions, RootCommandDefinition } from './types.ts';