+++ packages/cliapp/src/declarative/option/array.ts
@@ -1,0 +1,7 @@
+import { BaseOption } from './base.ts';
+
+export class ArrayOption extends BaseOption<string[]> {
+  parse(value: string): string[] {
+    return value.split(',').map((s) => s.trim());
+  }
+}