+++ packages/cliapp/src/declarative/option/mod.ts
@@ -1,0 +1,8 @@
+export { ArrayOption as Array } from './array.ts';
+export { BaseOption as Base } from './base.ts';
+export { BooleanOption as Boolean } from './boolean.ts';
+export { option } from './consts.ts';
+export { DateOption as Date } from './date.ts';
+export { NumberOption as Number } from './number.ts';
+export { PathOption as Path } from './path.ts';
+export { StringOption as String } from './string.ts';