+++ packages/cliapp/src/declarative/option/string.ts
@@ -1,0 +1,7 @@
+import { BaseOption } from './base.ts';
+
+export class StringOption extends BaseOption<string> {
+  parse(value: string): string {
+    return value;
+  }
+}