+++ packages/cliapp/src/types.ts
@@ -6,6 +6,7 @@
  */
 
 import type * as Log from '@epdoc/logger';
+import type { IBaseCtx } from './declarative/types.ts';
 import type { Console } from '@epdoc/msgbuilder';
 
 /**
@@ -18,6 +19,10 @@ export type DenoPkg = {
   author?: { name?: string; email?: string };
   workspace?: string[];
   license?: string;
+  repository?: {
+    type: string;
+    url: string;
+  };
 };
 
 /**
@@ -36,7 +41,7 @@ export type Logger<M extends MsgBuilder = MsgBuilder> = Log.Std.Logger<M>;
 export interface ICtx<
   M extends MsgBuilder = MsgBuilder,
   L extends Logger<M> = Logger<M>,
-> {
+> extends IBaseCtx {
   log: L;
   logMgr: Log.Mgr<M>;
   dryRun: boolean;